requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
lightning==2.5.2
lightning-utilities==0.15.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
yarl==1.20.1
